Mise a jour des Classe pour la base de données.  Question en erreur et commenter du faite que la table media n'est pas Terminée (CursorToItem).

+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/data/TeamSQLiteAdapter.java
@@ -94,7 +94,7 @@ public class TeamSQLiteAdapter {
      * @param id
      * @return Team
      */
-    public Team getTeam(long id){
+    public Team getTeam(int id){
 
         String[] cols = {COL_ID, COL_ID_SERVER, COL_NAME, COL_UPDATED_AT};
         String whereClausesSelect = COL_ID + "= ?";
